% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module_loadings.R
\name{plot_module_loadings}
\alias{plot_module_loadings}
\title{Plot feature loadings for a module eigengene}
\usage{
plot_module_loadings(loadings, module)
}
\arguments{
\item{loadings}{Named list of PCA loadings per module,
as returned by compute_eigengenes().}

\item{module}{Character. Name of the module to plot.}
}
\value{
A ggplot object.
}
\description{
Visualises PCA loadings for features within a given module,
highlighting features that contribute more than expected by chance.
}
