% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_clustering_bounds.R
\name{suggest_clustering_bounds}
\alias{suggest_clustering_bounds}
\title{Suggest clustering bounds for feature modularisation}
\usage{
suggest_clustering_bounds(n_features)
}
\arguments{
\item{n_features}{Integer. Number of features to cluster.}
}
\value{
A list with suggested clustering parameters.
}
\description{
Provides data-driven default bounds for the number of clusters
and minimum module size, based on feature dimensionality.
}
